# api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: User Management APIs
host: 'localhost:10004'
basePath: /rbac
schemes:
- http
consumes:
- multipart/form-data
- application/json
produces:
- application/json
- text/plain
paths:
...
  '/authType/{userName}':
    x-swagger-router-controller: controller
    get:
      description: Retrieve authType of a User
      operationId: authType
      parameters:
      - name: authorization
        in: header
        type: string
        description: The JWT token for req. validation
      - name: userName
        in: path
        type: string
        required: true
        description: UserName of the object to be updated
      responses:
        '200':
          description: List of the entites
        '400':
          description: Bad parameters
        '404':
          description: No results to list with the given parameter set.
        '500':
          description: Internal server error 
...